requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argcomplete==3.6.3
attrs==25.4.0
beautifulsoup4==4.14.3
blivet==3.12.1
blivet-gui==2.6.0
boto3==1.42.39
botocore==1.42.39
Brlapi==0.8.7
browser-cookie3==0.20.1
certifi==2025.7.9
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.1.7
cockpit @ file:///builddir/build/BUILD/cockpit-354-build/cockpit-354/tmp/wheel/cockpit-354-py3-none-any.whl#sha256=1d4fb1517c54d5d11d999be3583ea5fbf3d3d011cb789fb717daae0672c38816
crypt_r==3.13.1
cupshelpers==1.0
dasbus==1.7
dbus-python==1.4.0
dbus_next==0.2.3
distro==1.9.0
dnf==4.24.0
fastapi==0.129.0
fedora-third-party==0.10
file-magic==0.4.0
FVS==0.3.4
gemini-webapi==1.19.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
hyperframe==6.1.0
icoextract==0.1.6
idna==3.10
initial-setup==0.3.101
iso639==0.1.4
jeepney==0.9.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langtable==0.0.69
libcomps==0.1.23
libdnf==0.75.0
libdnf5==5.2.17.0
loguru==0.7.3
louis==3.33.0
lxml==6.0.1
lz4==4.4.5
Markdown==3.10
nftables==0.1
olefile==0.47
orjson==3.11.7
packaging==25.0
pathvalidate==3.3.1
patool==4.0.2
pefile==2024.8.26
perf==0.1
pexpect==4.9.0
pid==3.0.4
pillow==11.3.0
ply==3.11
productmd==1.50
psutil==7.0.0
ptyprocess==0.7.0
pwquality==1.4.5
PyAudio==0.2.13
pycairo==1.28.0
pycparser==2.22
pycryptodomex==3.23.0
pycups==2.0.4
pycurl==7.45.6
pydantic==2.12.5
pydantic_core==2.41.5
pyenchant==3.2.2
pygdbmi==0.11.0.0
PyGObject==3.54.5
pyinotify==0.9.6
pykickstart @ file:///builddir/build/BUILD/pykickstart-3.66-build/pykickstart-3.66
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-linux-procfs==0.7.3
python-meh==0.52
pyudev==0.24.3
pyxdg==0.28
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.18.7-build/kernel-6.18.7/linux-6.18.7-200.fc43.x86_64/tools/net/ynl
RapidFuzz==3.12.2
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
requests-file==2.0.0
requests-ftp==0.3.1
rpds-py==0.27.0
rpm==6.0.1
s3transfer==0.16.0
selinux @ file:///builddir/build/BUILD/libselinux-3.9-build/libselinux-3.9/src
sentry-sdk==2.35.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.9-build/selinux-3.9/python/sepolicy
setools==4.6.0
setroubleshoot @ file:///builddir/build/BUILD/setroubleshoot-3.3.36-build/setroubleshoot-3.3.36/src
setuptools==78.1.1
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sos==4.10.2
soupsieve==2.8
starlette==0.52.1
systemd-python==235
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
vkbasalt-cli==3.1.1.post2
xkbregistry==0.3
